Move check of removed configs to setConfWithCheck
